# PyPIRepository: cache return value of get_dependencies