This improves performance, especially for VCS editable dependencies.

Fixes: https://github.com/nvie/pip-tools/issues/159

Before:
pip-compile -v requirements/base.in  5,33s user 0,83s system 38% cpu 16,121 total

After:
pip-compile -v requirements/base.in  3,44s user 0,40s system 45% cpu 8,346 total

I have only a handful of editables, and it took 3 rounds to become stable.