@blueyed 2015-06-17
The test failures are unrelated: see #167 for fixes.