@blueyed 2015-06-17
This is included in https://github.com/nvie/pip-tools/pull/169 also.